chore:move json helpers to utils folder
